Move fusions.* and fusion_emitter.* to backends/gpu/codegen directory.
PiperOrigin-RevId: 716625085
