+++ xla/backends/gpu/codegen/BUILD
@@ -20,6 +20,7 @@ cc_library(
     srcs = ["copy.cc"],
     hdrs = ["copy.h"],
     deps = [
+        ":fusion_emitter",
         "//xla:shape_util",
         "//xla/backends/gpu/runtime:copy_thunk",
         "//xla/backends/gpu/runtime:thunk",
@@ -28,7 +29,6 @@ cc_library(
         "//xla/service:buffer_assignment",
         "//xla/service/gpu:hlo_fusion_analysis",
         "//xla/service/gpu:ir_emitter_context",
-        "//xla/service/gpu/fusions:fusion_emitter",
         "@com_google_absl//absl/status",
         "@com_google_absl//absl/status:statusor",
         "@tsl//tsl/platform:errors",
@@ -42,6 +42,7 @@ cc_library(
     hdrs = ["cudnn.h"],
     local_defines = if_cuda_is_configured(["GOOGLE_CUDA=1"]),
     deps = [
+        ":fusion_emitter",
         "//xla/backends/gpu/runtime:cudnn_thunk",
         "//xla/backends/gpu/runtime:thunk",
         "//xla/hlo/ir:hlo",
@@ -49,7 +50,6 @@ cc_library(
         "//xla/service/gpu:ir_emitter_context",
         "//xla/service/gpu:kernel_arguments",
         "//xla/service/gpu:kernel_reuse_cache",
-        "//xla/service/gpu/fusions:fusion_emitter",
         "@com_google_absl//absl/log",
         "@com_google_absl//absl/status:statusor",
         "@tsl//tsl/platform:statusor",
@@ -103,6 +103,7 @@ cc_library(
     srcs = ["custom.cc"],
     hdrs = ["custom.h"],
     deps = [
+        ":fusion_emitter",
         "//xla:literal",
         "//xla:literal_util",
         "//xla:shape_util",
@@ -135,7 +136,6 @@ cc_library(
         "//xla/service/gpu:kernel_arguments",
         "//xla/service/gpu:matmul_utils",
         "//xla/service/gpu:stream_executor_util",
-        "//xla/service/gpu/fusions:fusion_emitter",
         "//xla/service/gpu/kernels:custom_kernel",
         "//xla/service/gpu/kernels:custom_kernel_fusion",
         "//xla/stream_executor:stream",
@@ -205,4 +205,77 @@ xla_test(
     ]) + if_rocm_is_configured([
         "@local_config_rocm//rocm:rocm_headers",
     ]),
+)
+
+cc_library(
+    name = "fusion_emitter",
+    srcs = ["fusion_emitter.cc"],
+    hdrs = ["fusion_emitter.h"],
+    visibility = [
+        "//xla/backends/gpu/codegen:__subpackages__",
+        "//xla/service/gpu:__subpackages__",
+    ],
+    deps = [
+        "//xla:shape_util",
+        "//xla:status_macros",
+        "//xla:util",
+        "//xla/backends/gpu/runtime:kernel_thunk",
+        "//xla/backends/gpu/runtime:thunk",
+        "//xla/hlo/analysis:indexing_analysis",
+        "//xla/hlo/ir:hlo",
+        "//xla/service/gpu:ir_emitter_context",
+        "//xla/service/gpu:kernel_arguments",
+        "//xla/service/gpu:kernel_reuse_cache",
+        "//xla/service/gpu:launch_dimensions",
+        "//xla/service/gpu:target_util",
+        "//xla/service/llvm_ir:ir_array",
+        "//xla/service/llvm_ir:llvm_util",
+        "//xla/stream_executor:device_description",
+        "@com_google_absl//absl/log",
+        "@com_google_absl//absl/log:check",
+        "@com_google_absl//absl/status",
+        "@com_google_absl//absl/status:statusor",
+        "@com_google_absl//absl/strings",
+        "@com_google_absl//absl/types:span",
+        "@llvm-project//llvm:Support",
+        "@llvm-project//llvm:TargetParser",
+        "@llvm-project//llvm:ir_headers",
+        "@llvm-project//mlir:IR",
+        "@tsl//tsl/platform:errors",
+        "@tsl//tsl/platform:statusor",
+    ],
+)
+
+cc_library(
+    name = "fusions",
+    srcs = ["fusions.cc"],
+    hdrs = ["fusions.h"],
+    visibility = [
+        "//xla/backends/gpu/codegen/triton:__subpackages__",
+        "//xla/service/gpu:__subpackages__",
+    ],
+    deps = [
+        ":copy",
+        ":cudnn",
+        ":custom",
+        ":fusion_emitter",
+        "//xla:shape_util",
+        "//xla/backends/gpu/codegen/emitters:concatenate",
+        "//xla/backends/gpu/codegen/emitters:in_place_dynamic_update_slice",
+        "//xla/backends/gpu/codegen/emitters:input_slices",
+        "//xla/backends/gpu/codegen/emitters:loop",
+        "//xla/backends/gpu/codegen/emitters:reduction",
+        "//xla/backends/gpu/codegen/emitters:scatter",
+        "//xla/backends/gpu/codegen/emitters:transpose",
+        "//xla/backends/gpu/codegen/triton:fusion",
+        "//xla/hlo/ir:hlo",
+        "//xla/hlo/utils:hlo_traversal",
+        "//xla/service:buffer_assignment",
+        "//xla/service/gpu:backend_configs_cc",
+        "//xla/service/gpu:hlo_fusion_analysis",
+        "//xla/service/gpu:ir_emission_utils",
+        "@com_google_absl//absl/algorithm:container",
+        "@com_google_absl//absl/status:statusor",
+        "@com_google_absl//absl/strings",
+    ],
 )
