+++ xla/backends/gpu/codegen/custom.h
@@ -16,8 +16,8 @@ limitations under the License.
 #define XLA_BACKENDS_GPU_CODEGEN_CUSTOM_H_
 
 #include "absl/status/statusor.h"
+#include "xla/backends/gpu/codegen/fusion_emitter.h"
 #include "xla/hlo/ir/hlo_instructions.h"
-#include "xla/service/gpu/fusions/fusion_emitter.h"
 #include "xla/service/gpu/hlo_fusion_analysis.h"
 #include "xla/service/gpu/ir_emitter_context.h"
