+++ xla/backends/gpu/codegen/fusion_emitter.cc
@@ -12,7 +12,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
-#include "xla/service/gpu/fusions/fusion_emitter.h"
+#include "xla/backends/gpu/codegen/fusion_emitter.h"
 
 #include <array>
 #include <cstddef>
