+++ xla/backends/gpu/codegen/fusion_emitter.h
@@ -12,8 +12,8 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
-#ifndef XLA_SERVICE_GPU_FUSIONS_FUSION_EMITTER_H_
-#define XLA_SERVICE_GPU_FUSIONS_FUSION_EMITTER_H_
+#ifndef XLA_BACKENDS_GPU_CODEGEN_FUSION_EMITTER_H_
+#define XLA_BACKENDS_GPU_CODEGEN_FUSION_EMITTER_H_
 
 #include <array>
 #include <cstddef>
@@ -132,4 +132,4 @@ absl::Status AnnotateKernelLaunchDimensions(
 }  // namespace gpu
 }  // namespace xla
 
-#endif  // XLA_SERVICE_GPU_FUSIONS_FUSION_EMITTER_H_
+#endif  // XLA_BACKENDS_GPU_CODEGEN_FUSION_EMITTER_H_
