+++ xla/backends/gpu/codegen/triton/BUILD
@@ -27,12 +27,16 @@ cc_library(
     name = "fusion",
     srcs = ["fusion.cc"],
     hdrs = ["fusion.h"],
-    visibility = ["//xla/service/gpu:__subpackages__"],
+    visibility = [
+        "//xla/backends/gpu/codegen:__subpackages__",
+        "//xla/service/gpu:__subpackages__",
+    ],
     deps = [
         ":fusion_emitter",
         ":fusion_emitter_legacy_matmul",
         "//xla:shape_util",
         "//xla:status_macros",
+        "//xla/backends/gpu/codegen:fusion_emitter",
         "//xla/backends/gpu/runtime:kernel_thunk",
         "//xla/hlo/ir:hlo",
         "//xla/hlo/utils:hlo_traversal",
@@ -45,7 +49,6 @@ cc_library(
         "//xla/service/gpu:launch_dimensions",
         "//xla/service/gpu:matmul_utils",
         "//xla/service/gpu:triton_fusion_analysis",
-        "//xla/service/gpu/fusions:fusion_emitter",
         "//xla/service/gpu/model:tiled_hlo_instruction_or_computation",
         "//xla/service/llvm_ir:ir_array",
         "//xla/service/llvm_ir:llvm_util",
@@ -71,11 +74,11 @@ xla_cc_test(
     tags = ["gpu"],
     deps = [
         ":fusion",
+        "//xla/backends/gpu/codegen:fusion_emitter",
+        "//xla/backends/gpu/codegen:fusions",
         "//xla/hlo/ir:hlo",
         "//xla/service/gpu:gpu_device_info_for_tests",
         "//xla/service/gpu:hlo_fusion_analysis",
-        "//xla/service/gpu/fusions",
-        "//xla/service/gpu/fusions:fusion_emitter",
         "//xla/stream_executor:device_description",
         "//xla/tests:hlo_test_base",
         "//xla/tests:xla_internal_test_main",
