+++ xla/service/gpu/fusions/tools/BUILD
@@ -12,13 +12,13 @@ cc_library(
     hdrs = ["test_lib.h"],
     deps = [
         "//xla:status_macros",
+        "//xla/backends/gpu/codegen:fusions",
         "//xla/backends/gpu/codegen/emitters:emitter_base",
         "//xla/backends/gpu/codegen/emitters/ir:xla_gpu",
         "//xla/hlo/ir:hlo",
         "//xla/mlir_hlo",
         "//xla/service/gpu:gpu_device_info_for_tests",
         "//xla/service/gpu:hlo_fusion_analysis",
-        "//xla/service/gpu/fusions",
         "//xla/stream_executor:device_description",
         "//xla/tools:hlo_module_loader",
         "@com_google_absl//absl/algorithm:container",
